requirements.txt
# Core API
fastapi>=0.111
uvicorn[standard]>=0.30
pydantic>=2.7

# HTTP + env
python-dotenv>=1.0
requests>=2.31
httpx==0.27.2

# Telegram
python-telegram-bot>=22,<23

# для RAG
langchain-community>=0.2.9
langchain-huggingface>=0.0.3
sentence-transformers>=2.6

# Chroma
langchain-chroma>=0.1.0
chromadb>=1.0.20

# Qdrant
langchain-qdrant>=0.1.0
qdrant-client>=1.9
langchain-huggingface>=0.0.3


# tests
pytest>=8.0

# не забыть добавить CUDA под пакет torch в будущем, пока на CPU

